IpcDetection/backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
pandas==2.0.0
joblib==1.3.0
scikit-learn==1.3.0
Pillow==10.0.0
pytesseract==0.3.10
matplotlib==3.8.0
seaborn==0.12.0
